refactoring, separated story functions into controller files
